Remove the legacy Thinger.
